packages/fkweb/pyproject.toml
[project]
name = "fkweb"
version = "0.1.0"
description = "Frikanalen's Django-based backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url==2.3.0",
    "django==5.1.7",
    "django-cors-headers==4.7.0",
    "django-csp==3.8",
    "django-durationfield==0.5.5",
    "django-filter==25.1",
    "django-model-utils==5.0.0",
    "django-phonenumber-field[phonenumbers]==8.0.0",
    "djangorestframework==3.16.0",
    "djangorestframework-camel-case==1.4.2",
    "drf-spectacular==0.28.0",
    "gunicorn==23.0.0",
    "markdown==3.7",
    "psycopg2-binary>=2.9.10",
    "python-dateutil==2.9.0.post0",
    "python-memcached==1.62",
    "requests==2.32.3",
]

[tool.ruff]
line-length = 100
